Confusing 'both' word deleted.
I guess, first the requirement was intended to be to preserve Both the
sorting order and the rating filter, but then it was changed.

The word 'both' stayed still. So I deleted it to make the sentence
less ambiguous.

Please, let me know, if this is my English, and I just get wrong
what was written!